AI-Agents/1. Chat WIth Your Data/requirements.txt
streamlit
pandas
matplotlib
langchain
python-dotenv
openai
